requirements.txt
fastapi==0.109.2
pydantic==2.6.1
openai==1.12.0
python-dotenv==1.0.1
uvicorn==0.27.1
supabase==2.4.3
